+++ CHANGE_LOG.md
@@ -1,4 +1,11 @@
-# 0.0.1 -- fuel-core
+# 0.0.1 -- new contracts
+
+1. Second testnet up with real finalization and submission delay
+2. Lambda privacy adjustments related to IP
+3. Further withdrawal wallet feature activation
+4. More testing
+
+# 0.0.1 -- launch
 
 1. Basic testing
 2. Basic docs